DevSB16.cpp: Crash fix.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@55074 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Audio/DevSB16.cpp
@@ -850,7 +850,7 @@ static void sb16HandleCommand(PSB16STATE pThis, uint8_t cmd)
 
             case 0xe3:
             {
-                for (size_t i = sizeof (e3) - 1; i >= 0; --i)
+                for (int i = sizeof (e3) - 1; i >= 0; --i)
                     dsp_out_data(pThis, e3[i]);
 
                 break;
